+++ frontend/src/pages/_document.tsx
@@ -1,11 +1,13 @@
 'use client'
 
 import { Html, Head, Main, NextScript } from "next/document";
-import Home from ".";
 import React from 'react';
 import { Nav } from "@/components/Nav";
+import Home from ".";
+import { getStaticProps } from "./api/usps_oauth";
 
-export default function Document() {
+
+export default async function Document() {
   return (
     <Html lang="en">
       <Head>
@@ -14,7 +16,7 @@ export default function Document() {
       <body className="antialiased">
         <Main />
         <NextScript />
-        <Home />
+        <Home token={(await getStaticProps()).props.token} />
       </body>
     </Html>
   );
